% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxInverseLink.R
\name{approxInverseLink_lcmm}
\alias{approxInverseLink_lcmm}
\title{Numeric Inverse Link Function of \link[lcmm]{lcmm} Object}
\usage{
approxInverseLink_lcmm(object, ...)
}
\arguments{
\item{object}{an \link[lcmm]{lcmm} object}

\item{...}{other parameters of \link[stats]{approxfun}}
}
\description{
To numerically retrieve the inverse link \emph{function} from a latent class mixed-effect model.
}
\note{
See section \strong{Value} of \link[lcmm]{lcmm}.

See \code{lcmm:::summary.lcmm} very carefully!

Functions, see Page 4 of \url{https://www.jstatsoft.org/article/view/v078i02}
\code{lcmm:::.Contlcmm}, then Fortran code \code{C_hetmixcont}, look for output \code{transfY};
actually in subroutine \code{transfo_estimee}.

I am totally confused, and cannot reproduce the results. So I finally resort to brutal force!!

\code{lcmm:::plot.lcmm} with option 'linkfunction' -> \code{lcmm:::.plotlinkfunction},
see \code{plot(object, which = 'link')}.
}
\examples{
# to validate
validate_inverseLink_lcmm = function(object) {
  marker <- object$estimlink[,1L]
  transfY <- object$estimlink[,2L]
  my_marker <- approxInverseLink_lcmm(object)(transfY)
  range(marker - my_marker)
}
validate_inverseLink_lcmm(m20)
validate_inverseLink_lcmm(m21)

}
