% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxlinkinv.R
\name{approxlinkinv_lcmm}
\alias{approxlinkinv_lcmm}
\title{Inverse Link Function of \link[lcmm]{lcmm} Object, via Linear Interpolation}
\usage{
approxlinkinv_lcmm(object, ...)
}
\arguments{
\item{object}{an \link[lcmm]{lcmm} object}

\item{...}{other parameters of function \link[stats]{approxfun}}
}
\description{
Inverse link \emph{function} from a latent class mixed-effect model, via linear interpolation.
}
\note{
See section \strong{Value} of \link[lcmm]{lcmm}.

See function \code{lcmm:::summary.lcmm()} very carefully!

Functions, see Page 4 of \url{https://www.jstatsoft.org/article/view/v078i02}
\code{lcmm:::.Contlcmm}, then Fortran code \code{C_hetmixcont}, look for output \code{transfY};
actually in subroutine \code{transfo_estimee}.

I am totally confused, and cannot reproduce the results. So I finally resort to brutal force!!

\code{lcmm:::plot.lcmm()} with option 'linkfunction' -> \code{lcmm:::.plotlinkfunction},
see \code{plot(object, which = 'link')}.
}
\keyword{internal}
