% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot_lcmm}
\alias{autoplot_lcmm}
\alias{autoplot.lcmm}
\alias{autolayer.lcmm}
\title{Plot \link[lcmm]{lcmm} Object}
\usage{
\method{autoplot}{lcmm}(object, ...)

\method{autolayer}{lcmm}(object, type = c("original", "transformed"), ...)
}
\arguments{
\item{object}{\link[lcmm]{lcmm} object}

\item{...}{..}

\item{type}{\link[base]{character} scalar,
\code{'original'} (default) or \code{'transformed'}}
}
\description{
..
}
\examples{
autoplot(m20)
autoplot(m20, type = 'trans')
autoplot(m21)
autoplot(m21, type = 'trans')

}
