% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lcmm.R
\name{formula.lcmm}
\alias{formula.lcmm}
\title{Model Formula of \link[lcmm]{lcmm} Object}
\usage{
\method{formula}{lcmm}(x, ...)
}
\arguments{
\item{x}{\link[lcmm]{lcmm} object}

\item{...}{additional parameters, currently of no use}
}
\description{
Model Formula of \link[lcmm]{lcmm} Object
}
\examples{
formula(m20)
}
