% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData.lcmm}
\alias{getData.lcmm}
\title{Get Data from \link[lcmm]{lcmm} Object}
\usage{
\method{getData}{lcmm}(object)
}
\arguments{
\item{object}{an \link[lcmm]{lcmm} object}
}
\value{
Function \code{\link[=getData.lcmm]{getData.lcmm()}} returns a \link[base]{data.frame}.
}
\description{
To obtain the input data for latent class mixed-effect model analysis.
}
\details{
The input \link[lcmm]{lcmm} \code{object} must have \code{returndata = TRUE} in its call.
Otherwise, an error is returned to tell the end user to
re-run \link[lcmm]{lcmm} call with \code{returndata = TRUE}.

\verb{$na.action} blah blah.
See argument \code{na.action} of \link[lcmm]{lcmm},
only \link[stats]{na.omit} and \link[stats]{na.fail} allowed, for now.
}
\note{
Be aware of the name clash with \code{?lcmm::data_lcmm}, which is a \link[base]{data.frame}.
}
\examples{
head(getData.lcmm(m20))

}
