% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{lcmm.tzh-package}
\alias{lcmm.tzh}
\alias{lcmm.tzh-package}
\title{lcmm.tzh: Extending lcmm Package}
\description{
To extend lcmm package.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

}
\keyword{internal}
