% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmm_B1.R
\name{lcmm_B1}
\alias{lcmm_B1}
\title{\link[lcmm]{lcmm} using Initial Value based on \code{ng = 1L}}
\usage{
lcmm_B1(fixed, ...)
}
\arguments{
\item{fixed}{see \link[lcmm]{lcmm}}

\item{...}{..}
}
\description{
..
}
\examples{
data(data_lcmm, package = 'lcmm')
m = lcmm_B1(Ydep2~Time, subject='ID',
 idiag=TRUE, data=data_lcmm, link='beta', ng = 2L,
 verbose=FALSE, returndata=TRUE)

}
