% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_doc.R
\docType{data}
\name{lcmm_examples}
\alias{lcmm_examples}
\alias{m20}
\alias{m21}
\title{\link[lcmm]{lcmm} Examples}
\format{
..

An object of class \code{lcmm} of length 38.
}
\usage{
m20

m21
}
\description{
..
}
\references{
\url{https://mcfromnz.wordpress.com/2011/10/02/latent-class-mixture-modeling-with-graphics/}
}
\keyword{datasets}
