% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmm_find_ng.R
\name{lcmm_find_ng}
\alias{lcmm_find_ng}
\title{\link[lcmm]{lcmm} with Optimal Number of Latent Classes}
\usage{
lcmm_find_ng(fixed, criterion = BIC, ...)
}
\arguments{
\item{fixed}{see \link[lcmm]{lcmm}}

\item{criterion}{..}

\item{...}{..}
}
\description{
..
}
\examples{
data(data_lcmm, package = 'lcmm')
m = lcmm_find_ng(Ydep2~Time, subject='ID',
 idiag=TRUE, data=data_lcmm, link='beta',
 verbose=FALSE, returndata=TRUE)
}
\keyword{internal}
