% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lcmm.R
\name{logLik.lcmm}
\alias{logLik.lcmm}
\title{Log-Likelihood of \link[lcmm]{lcmm} Object}
\usage{
\method{logLik}{lcmm}(object, ...)
}
\arguments{
\item{object}{\link[lcmm]{lcmm} object}

\item{...}{additional parameters}
}
\value{
Function \code{\link[=logLik.lcmm]{logLik.lcmm()}} returns a \link[stats]{logLik} object.
}
\description{
..
}
\note{
\link[lcmm]{lcmm} object contains elements \verb{$loglik}, \verb{$AIC} and \verb{$BIC}.
}
\examples{
m20 |> logLik()
m20 |> AIC() |> all.equal.numeric(current = m20$AIC) |> stopifnot()
m20 |> BIC() |> all.equal.numeric(current = m20$BIC) |> stopifnot()
}
\keyword{internal}
