% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lcmm.R
\name{logLik_lcmm}
\alias{logLik_lcmm}
\alias{logLik.lcmm}
\alias{AIC.lcmm}
\alias{BIC.lcmm}
\title{Log-Likelihood, AIC and BIC of \link[lcmm]{lcmm} Object}
\usage{
\method{logLik}{lcmm}(object, ...)

\method{AIC}{lcmm}(object, ...)

\method{BIC}{lcmm}(object, ...)
}
\arguments{
\item{object}{\link[lcmm]{lcmm} object}

\item{...}{additional parameters}
}
\value{
Function \code{\link[=logLik.lcmm]{logLik.lcmm()}} returns a \link[stats]{logLik} object.

Functions \code{\link[=AIC.lcmm]{AIC.lcmm()}} and \code{\link[=BIC.lcmm]{BIC.lcmm()}} return \link[base]{numeric} scalars.
}
\description{
..
}
\details{
\link[lcmm]{lcmm} object contains elements \verb{$loglik}, \verb{$AIC} and \verb{$BIC}.
}
\examples{
logLik(m20)
AIC(m20)
BIC(m20)
}
