% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lcmm.R
\name{nobs.lcmm}
\alias{nobs.lcmm}
\title{Number of Subjects in \link[lcmm]{lcmm} Object}
\usage{
\method{nobs}{lcmm}(object, ...)
}
\arguments{
\item{object}{an \link[lcmm]{lcmm} object}

\item{...}{additional parameters}
}
\value{
Function \code{\link[=nobs.lcmm]{nobs.lcmm()}} returns an \link[base]{integer} scalar.
}
\description{
Number of Subjects in \link[lcmm]{lcmm} Object
}
\details{
This is clearly documented in \link[lcmm]{lcmm}.

Method dispatch to S3 generic \link[stats]{nobs} is required in S3 generic \link[stats]{BIC}.
The end user should use function \code{\link[=nobsText.lcmm]{nobsText.lcmm()}}.
}
\examples{
m20 |> nobs()
}
