% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsClass.R
\name{obsClass}
\alias{obsClass}
\alias{obsClass.lcmm}
\title{Identifier of Latent Classes from \link[lcmm]{lcmm} Object}
\usage{
obsClass(object)

\method{obsClass}{lcmm}(object)
}
\arguments{
\item{object}{an \link[lcmm]{lcmm} object}
}
\value{
Function \link{obsClass.lcmm} returns a \link[base]{factor}.
}
\description{
To obtain the identifier of latent classes from an latent class mixed-effect model.
}
\note{
See \strong{Value} section of \link[lcmm]{lcmm}.

Name clash \code{methods::getClass}.

Function \link[lcmm]{xclass} requires two models fitted to a same \code{data}.
}
\examples{
data(data_lcmm, package = 'lcmm')
dim(data_lcmm)
table(m20$pprob$class) # per subject
table(obsClass(m20)) # per observation
}
