% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_ng.R
\name{reduce_ng_lcmm}
\alias{reduce_ng_lcmm}
\title{Reduce Number of Latent Classes from \link[lcmm]{lcmm} Object}
\usage{
reduce_ng_lcmm(object, criterion = BIC, ...)
}
\arguments{
\item{object}{an \link[lcmm]{lcmm} object}

\item{criterion}{\link[base]{function}, criterion of elimination.
Either \link[stats]{AIC} or \link[stats]{BIC}.}

\item{...}{potential parameters, currently not in use}
}
\description{
To reduce number of latent classes from \link[lcmm]{lcmm} object.
}
\details{
For all three methods (\link[stats]{logLik}, \link[stats]{AIC} and \link[stats]{BIC}),
the comparison is always with the original model.

This is different from \link[MASS]{stepAIC}, in which the comparison is with the previous round.
}
\examples{
reduce_ng_lcmm(m20)

}
