% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lcmm.R
\name{S3_lcmm}
\alias{S3_lcmm}
\alias{nobsText.lcmm}
\alias{desc_.lcmm}
\title{Additional S3 Method Dispatches for \link[lcmm]{lcmm} Object}
\usage{
nobsText.lcmm(x)

desc_.lcmm(x)
}
\arguments{
\item{x}{\link[lcmm]{lcmm} object}
}
\value{
Function \code{\link[=nobsText.lcmm]{nobsText.lcmm()}} returns a \link[base]{character} scalar.
}
\description{
Additional S3 Method Dispatches for \link[lcmm]{lcmm} Object
}
\note{
See \code{lcmm:::summary.lcmm} for details.
}
\examples{
nobsText.lcmm(m20)
}
\keyword{internal}
