% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lcmm.R
\name{md_.lcmm}
\alias{md_.lcmm}
\title{R Markdown Lines for \link[lcmm]{lcmm} Object}
\usage{
\method{md_}{lcmm}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link[lcmm]{lcmm} Object
}
\examples{
list(
 'model 1' = m20,
 'model 2' = m21
) |> fastmd::render2html(file = 'lcmm')
}
\keyword{internal}
